+++ data-raw/data-raw.R
@@ -1,3 +1,5 @@
+suppressPackageStartupMessages(source('data-raw/drug_utilities.R'))
+suppressPackageStartupMessages(source('data-raw/biomarker_utilities.R'))
 
 ## get metadata from metadata_pharma_oncox.xlsx
 metadata <- list()
@@ -25,8 +27,6 @@ chembl_pubchem_datestamp <- '20220906'
 lgr::lgr$appenders$console$set_layout(
   lgr::LayoutFormat$new(timestamp_fmt = "%Y-%m-%d %T"))
 
-suppressPackageStartupMessages(source('data-raw/drug_utilities.R'))
-suppressPackageStartupMessages(source('data-raw/biomarker_utilities.R'))
 
 
 nci_ftp_base <- paste0("https://evs.nci.nih.gov/ftp1/NCI_Thesaurus/archive/",
@@ -43,6 +43,9 @@ gene_basic <- geneOncoX::get_basic(
   cache_dir = path_data_tmp_processed)
 gene_gencode <- geneOncoX::get_gencode(
   cache_dir = path_data_tmp_processed)
+gene_alias <- geneOncoX::get_alias(
+  cache_dir = path_data_tmp_processed
+)
 
 gene_info <- dplyr::bind_rows(
   dplyr::inner_join(
@@ -105,7 +108,7 @@ if (!file.exists(antineo_agents_local)) {
 ## Get all anticancer drugs, NCI thesaurus + DGIdb
 nci_antineo_all <- get_nci_drugs(
   nci_db_release = nci_db_release,
-  overwrite = T,
+  overwrite = F,
   path_data_raw = path_data_raw,
   path_data_processed = path_data_tmp_processed)
 
@@ -156,39 +159,29 @@ drug_index_map[['id2alias']] <- drug_aliases
 drug_index_map$id2synonym <- NULL
 
 compound_synonyms <- drug_index_map[['id2alias']] |>
-  dplyr::mutate(drugname_lc = tolower(alias)) |>
+  dplyr::mutate(alias_lc = tolower(alias)) |>
   dplyr::left_join(
     drug_index_map[['id2name']], by = "drug_id",
     multiple = "all") |>
   dplyr::left_join(
     dplyr::select(drug_index_map[['id2basic']],
                   drug_id, molecule_chembl_id),
-    by = "drug_id", multiple = "all")
+    by = "drug_id", multiple = "all") |>
+  dplyr::select(
+    drug_id,
+    drug_name,
+    alias,
+    alias_lc,
+    molecule_chembl_id
+  ) |>
+  dplyr::filter(alias != molecule_chembl_id) |>
+  dplyr::distinct()
+  
 
 rm(drug_df)
 
 
 ##################### BIOMARKERS ###############################
-
-civic_variant_summary <- 
-  paste0("data-raw/biomarkers/civic/variant_summary_",
-         package_datestamp,
-         ".tsv")
-civic_clinical_evidence <- 
-  paste0("data-raw/biomarkers/civic/clinical_evidence_summary_",
-         package_datestamp,
-         ".tsv")
-
-if(!file.exists(civic_variant_summary)){
-  download.file(
-    url = "https://civicdb.org/downloads/nightly/nightly-VariantSummaries.tsv",
-    destfile = civic_variant_summary)
-}
-if(!file.exists(civic_clinical_evidence)){
-  download.file(
-    url = "https://civicdb.org/downloads/nightly/nightly-ClinicalEvidenceSummaries.tsv",
-    destfile = civic_clinical_evidence)
-}
 
 # gene_info <- get_gene_info_ncbi(
 #   path_data_raw = file.path(here::here(), "data-raw"),
@@ -199,160 +192,24 @@ raw_biomarkers[['civic']] <-
   load_civic_biomarkers(
     compound_synonyms = compound_synonyms,
     datestamp = package_datestamp,
-    cache_dir = path_data_raw)
+    cache_dir = file.path(path_data_raw, "biomarkers"))
 raw_biomarkers[['cgi']] <- load_cgi_biomarkers(
   compound_synonyms = compound_synonyms,
-  cache_dir = path_data_raw)
-raw_biomarkers[['mitelmandb']] <- load_mitelman_db(cache_dir = path_data_raw)
-raw_biomarkers[['pmkb']] <- load_pmkb_biomarkers(
-  cache_dir = path_data_raw
-)
-custom_fusiondb <- load_custom_fusion_db() |>
-  dplyr::anti_join(raw_biomarkers[["mitelmandb"]], by = "variant")
+  cache_dir = file.path(path_data_raw, "biomarkers"))
+raw_biomarkers[['mitelmandb']] <- load_mitelman_db(
+  cache_dir = file.path(path_data_raw, "biomarkers"))
+# raw_biomarkers[['pmkb']] <- load_pmkb_biomarkers(
+#   cache_dir = path_data_raw
+# )
+raw_biomarkers[['custom_fusions']] <- load_custom_fusion_db()
 
-biomarkers_all <- do.call(rbind, raw_biomarkers)
-rownames(biomarkers_all) <- NULL
-biomarkers_all <- biomarkers_all |>
-  dplyr::bind_rows(custom_fusiondb) |>
-  dplyr::mutate(therapeutic_context =
-                  stringi::stri_enc_toascii(therapeutic_context)) |>
-  dplyr::mutate(cancer_type =
-                  stringi::stri_enc_toascii(cancer_type)) |>
-  dplyr::mutate(evidence_description =
-                  stringi::stri_enc_toascii(evidence_description))
+raw_biomarkers[['custom_fusions']]$variant <- raw_biomarkers[['custom_fusions']]$variant |>
+  dplyr::anti_join(raw_biomarkers[["mitelmandb"]][['variant']], by = "variant_alias")
 
-rm(raw_biomarkers)
-biomarkers_curated <- list()
-biomarkers_curated[['records']] <- biomarkers_all
-biomarkers_curated[['metadata']] <- metadata$biomarkers[1:4,]
-rm(biomarkers_all)
-rm(civic_clinical_evidence)
-rm(civic_variant_summary)
-
-ccle_drugs <- readr::read_csv(
-  file="data-raw/biomarkers/ccle_depmap/primary-screen-replicate-treatment-info.csv.gz",
-  show_col_types = F) |>
-  dplyr::filter(perturbation_type == "experimental_treatment") |>
-  dplyr::filter(!is.na(broad_id)) |>
-  dplyr::select(broad_id, name, target,
-                moa, indication) |>
-  dplyr::rename(compound_id = broad_id,
-                drugname = name,
-                drugtarget = target,
-                drug_moa = moa,
-                drug_indication = indication) |>
-  dplyr::mutate(drugname_lc = tolower(drugname)) |>
-  dplyr::distinct()
-
-ccle_predictive_features <- readr::read_csv(
-  file="data-raw/biomarkers/ccle_depmap/Repurposing_primary_predictability_results.csv.gz",
-  show_col_types = F) |>
-  dplyr::mutate(compound_id = stringr::str_replace(gene,"BRD:","")) |>
-  dplyr::semi_join(ccle_drugs, by = c("compound_id"))
-
-all_predictive_features <- data.frame()
-
-for(n in 1:nrow(ccle_predictive_features)){
-  row <- as.data.frame(ccle_predictive_features[n,])
-  compound_id <- stringr::str_replace(row$gene,"BRD:","")
-  model <- row$model
-  pearson <- row$pearson
-  best <- row$best
-
-  if(model != "Core_omics"){
-    next
-  }
-
-  for(i in 0:9){
-    col1 <- paste0("feature",i)
-    col2 <- paste0("feature",i,"_importance")
-    symbol <- NA
-    feature_type <- NA
-    feature_value <- NA
-    if(stringr::str_detect(row[col1],"(CN|MutHot|MutMis|RNAseq|MutDam)$")){
-      feature_properties <- stringr::str_split(
-        stringr::str_replace_all(as.character(row[col1]),"\\)|\\(",""),"_")[[1]]
-      if(length(feature_properties) == 3){
-        symbol <- feature_properties[1]
-        entrezgene <- feature_properties[2]
-        feature_value <- paste(symbol, entrezgene, sep = ";")
-        feature_type <- feature_properties[3]
-        if(feature_type == "CN"){
-          feature_type = "CopyNumber"
-        }
-        if(feature_type == "MutMmis"){
-          feature_type = "MissenseMutation"
-        }
-        if(feature_type == "MutDam"){
-          feature_type = "DamagingMutation"
-        }
-        if(feature_type == "MutHot"){
-          feature_type = "HotSpotMutation"
-        }
-      }
-    }else{
-      if(stringr::str_detect(row[col1],"Lin$")){
-        feature_value = stringr::str_replace(row[col1],"_Lin$","")
-        feature_type = "Lineage"
-      }
-      else if(stringr::str_detect(row[col1],"_Fusion$")){
-        fusion_candidate = stringr::str_replace(row[col1],"_Fusion$","")
-        if(stringr::str_count(fusion_candidate,"_") == 1){
-          feature_type = "Fusion"
-          feature_value <- fusion_candidate
-        }
-      }
-    }
-    importance <- round(as.numeric(row[col2]), digits = 7)
-
-    if(!is.na(feature_type)){
-      df <- data.frame(
-        'compound_id' = compound_id,
-        'model' = model,
-        'feature_type' = feature_type,
-        'feature_value' = feature_value,
-        'feature_importance' = importance,
-        'prediction_accuracy' = pearson,
-        stringsAsFactors = F
-      )
-
-      all_predictive_features <- all_predictive_features |>
-        dplyr::bind_rows(df)
-    }
-
-  }
-}
-
-rm(col1)
-rm(col2)
-rm(best)
-rm(feature_properties)
-rm(feature_type)
-rm(feature_value)
-rm(compound_id)
-rm(importance)
-rm(model)
-rm(fusion_candidate)
-rm(pearson)
-
-biomarkers_invitro <- list()
-biomarkers_invitro[['metadata']] <- metadata$biomarkers[5,]
-biomarkers_invitro[['records']] <- all_predictive_features |>
-  dplyr::inner_join(ccle_drugs, by = "compound_id", multiple = "all") |>
-  dplyr::inner_join(compound_synonyms,
-                    by = "drugname_lc", multiple = "all") |>
-  dplyr::select(-c(drugname_lc, drugtarget, drug_moa,
-                   drug_indication, drugname)) |>
-  dplyr::distinct() |>
-  dplyr::select(-alias) |>
-  dplyr::distinct()
-
-
-rm(ccle_drugs)
-rm(ccle_predictive_features)
-rm(all_predictive_features)
-rm(compound_synonyms)
-
+biomarkers <- list()
+biomarkers[['data']] <- raw_biomarkers
+biomarkers[['metadata']] <- rbind(metadata$biomarkers[1:2,], metadata$biomarkers[4,])
+#rm(biomarkers_all)
 
 ## upload to Google Drive
 version_bump <- paste0(
@@ -362,8 +219,7 @@ version_bump <- paste0(
 
 
 db <- list()
-db[['biomarkers_curated']] <- biomarkers_curated
-db[['biomarkers_invitro']] <- biomarkers_invitro
+db[['biomarkers']] <- biomarkers
 db[['drug_map_name']] <- list()
 db[['drug_map_name']][['records']] <- drug_index_map[['id2name']]
 db[['drug_map_target']] <- list()
@@ -380,17 +236,14 @@ db[['drug_map_alias']][['records']] <- drug_index_map[['id2alias']]
 gd_records <- list()
 db_id_ref <- data.frame()
 
-
-for(elem in c('biomarkers_curated',
-              'biomarkers_invitro',
+for(elem in c('biomarkers',
               'drug_map_name',
               'drug_map_target',
               'drug_map_indication',
               'drug_map_basic',
               'drug_map_alias')){
 
-  if(elem != "biomarkers_curated" & 
-     elem != "biomarkers_invitro"){
+  if(elem != "biomarkers"){
     db[[elem]][['metadata']] <- metadata$compounds
   }